Choose only one interface among those with same subnet
E.g., if both wired and wireless interface are connected to the same router
(and thus, are in the same subnet), only listen/reply for discovery on best
interface (wired).

If interfaces are connected to different subnets (e.g., wireless connected
to router, wired with a cross cable), nothing changes - listen on both
interfaces.